move: 예외 핸들링 패키지를 common 내부로 이동
