+++ src/main/java/roomescape/common/globalexception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.globalexception;
+package roomescape.common.globalexception;
 
 public record ErrorResponse(
     String message
