+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -2,7 +2,7 @@ package roomescape.reservation.domain;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import roomescape.globalexception.BadRequestException;
+import roomescape.common.globalexception.BadRequestException;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
