+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -3,8 +3,8 @@ package roomescape.reservation.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import roomescape.globalexception.BadRequestException;
-import roomescape.globalexception.ConflictException;
+import roomescape.common.globalexception.BadRequestException;
+import roomescape.common.globalexception.ConflictException;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
 import roomescape.reservation.ReservationMapper;
