+++ src/main/java/roomescape/reservationtime/repository/ReservationTimeRepositoryImpl.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.globalexception.NotFoundException;
+import roomescape.common.globalexception.NotFoundException;
 import roomescape.reservationtime.domain.ReservationTime;
 
 @Repository
