+++ src/main/java/roomescape/reservationtime/service/ReservationTimeService.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import roomescape.globalexception.BadRequestException;
-import roomescape.globalexception.ConflictException;
+import roomescape.common.globalexception.BadRequestException;
+import roomescape.common.globalexception.ConflictException;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservationtime.ReservationTimeMapper;
