+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -3,8 +3,8 @@ package roomescape.theme.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.globalexception.BadRequestException;
-import roomescape.globalexception.NotFoundException;
+import roomescape.common.globalexception.BadRequestException;
+import roomescape.common.globalexception.NotFoundException;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.dto.ThemeReqDto;
