+++ src/test/java/roomescape/reservationtime/service/ReservationTimeServiceTest.java
@@ -14,8 +14,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.globalexception.BadRequestException;
-import roomescape.globalexception.NotFoundException;
+import roomescape.common.globalexception.BadRequestException;
+import roomescape.common.globalexception.NotFoundException;
 import roomescape.reservation.domain.dto.ReservationReqDto;
 import roomescape.reservation.repository.ReservationRepositoryImpl;
 import roomescape.reservation.service.ReservationService;
